tests/wpt/web-platform-tests/webmessaging/with-options/invalid-args.tentative.html

<!doctype html>
<title>unknown parameter</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws(new TypeError, () => postMessage('', '/', {targetOrigin: 'http://example.com'}));
});
</script>
